# ta-api-master/swagger/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/cancel": {
      "post": {
        "operationId": "GenericCancelController_cancel",
        "parameters": [
          {
            "name": "Currency-Preference",
            "in": "header",
            "description": "Please enter preferred currency code. (3 digit currency code)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "language",
            "in": "header",
            "description": "Please enter preferred language. (e.g., english,french)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Ip-Address",
            "in": "header",
            "description": "Please enter IP address. (e.g., 127.0.0.1)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Api-version",
            "in": "header",
            "description": "Please enter API version.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "v1"
              ]
            }
          },
          {
            "name": "Club-Id",
            "in": "header",
            "description": "Please enter club id.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Device-Information",
            "in": "header",
            "description": "Please enter device-information id.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenericCancelDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Entity not found"
          },
          "422": {
            "description": "Unprocessable entity"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Generic cancel API (flight/hotel)",
        "tags": [
          "Cancel"
        ]
      }
    },
    "/cancellation-charges": {
      "post": {
        "operationId": "GenericCancelController_getCancellationCharges",
        "parameters": [
          {
            "name": "Currency-Preference",
            "in": "header",
            "description": "Please enter preferred currency code. (3 digit currency code)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "language",
            "in": "header",
            "description": "Please enter preferred language. (e.g., english,french)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Ip-Address",
            "in": "header",
            "description": "Please enter IP address. (e.g., 127.0.0.1)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Api-version",
            "in": "header",
            "description": "Please enter API version.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "v1"
              ]
            }
          },
          {
            "name": "Club-Id",
            "in": "header",
            "description": "Please enter club id.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Device-Information",
            "in": "header",
            "description": "Please enter device-information id.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenericGetCancellationChargesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Entity not found"
          },
          "422": {
            "description": "Unprocessable entity"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Generic cancellation charges (flight/hotel)",
        "tags": [
          "Cancel"
        ]
      }
    }
  },
  "info": {
    "title": "API SERVICE",
    "description": "## Travel API Documentation\n\nWelcome to the **API SERVICE** documentation. This guide provides comprehensive guidelines and detailed instructions on how to effectively integrate and utilize the Travel API.\n\n### Overview\nThis documentation includes various sections covering:\n- **Authentication Methods**: Learn how to authenticate using Bearer and Cookie authentication.\n- **Endpoint Descriptions**: Detailed information on each API endpoint.\n- **Usage Examples**: Practical examples to help you get started quickly.\n- **Best Practices**: Tips and recommendations for efficient implementation.\n\n### Purpose\nWhether you are a developer looking to incorporate travel-related functionalities or a business aiming to enhance your travel services, this documentation serves as a valuable resource to guide you through the process.\n\n### Getting Started\nTo get started, ensure you have the necessary authentication tokens and follow the endpoint descriptions to make your first API call.\n\nFor any questions or support, please refer to the contact section at the end of this documentation.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "auth"
      }
    },
    "schemas": {
...
      "SectorDto": {
        "type": "object",
        "properties": {
          "origin": {
            "type": "string",
            "description": "Origin airport code",
            "example": "DEL"
          },
          "destination": {
            "type": "string",
            "description": "Destination airport code",
            "example": "BOM"
          }
        },
        "required": [
          "origin",
          "destination"
        ]
      },
      "SupplierParamsDto": {
        "type": "object",
        "properties": {
          "remarks": {
            "type": "string",
            "description": "Remarks for cancellation",
            "example": "Customer requested cancellation via portal"
          },
          "cancellationType": {
            "type": "string",
            "description": "Cancellation type",
            "example": "Others",
            "enum": [
              "NotSet",
              "NoShow",
              "FlightCancelled",
              "Others"
            ]
          },
          "ticketIds": {
            "description": "Array of ticket IDs (for partial cancellation)",
            "example": [
              123456,
              123457
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "sectors": {
            "description": "Sectors for partial cancellation",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SectorDto"
            }
          },
          "releasePnr": {
            "type": "boolean",
            "description": "Whether to release PNR (for hold bookings only)",
            "example": false
          }
        }
      },
      "GenericCancelDto": {
        "type": "object",
        "properties": {
          "mode": {
            "type": "string",
            "description": "Cancellation mode",
            "enum": [
              "flight",
              "hotel"
            ],
            "example": "flight"
          },
          "bookingId": {
            "type": "number",
            "description": "Unique booking ID",
            "example": 123456
          },
          "requestType": {
            "type": "string",
            "description": "Request type",
            "example": "FullCancellation",
            "enum": [
              "FullCancellation",
              "PartialCancellation",
              "Reissuance",
              "NotSet"
            ]
          },
          "supplierParams": {
            "description": "Supplier-specific parameters (varies by provider)",
            "allOf": [
              {
                "$ref": "#/components/schemas/SupplierParamsDto"
              }
            ]
          }
        },
        "required": [
          "mode",
          "bookingId",
          "requestType"
        ]
      },
      "GenericGetCancellationChargesDto": {
        "type": "object",
        "properties": {
          "mode": {
            "type": "string",
            "description": "Cancellation mode",
            "enum": [
              "flight",
              "hotel"
            ],
            "example": "flight"
          },
          "bookingId": {
            "type": "number",
            "description": "Unique booking ID",
            "example": 123456
          },
          "requestType": {
            "type": "string",
            "description": "Request type",
            "example": "FullCancellation",
            "enum": [
              "FullCancellation",
              "PartialCancellation",
              "Reissuance"
            ]
          }
        },
        "required": [
          "mode",
          "bookingId",
          "requestType"
        ]
      }
    }
  }
}